+++ src/Factory/Encoder/JsonEncoderFactory.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Para\Factory\Encoder;
+
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+
+/**
+ * Class JsonEncoderFactory.
+ *
+ * @package Factory\Encoder
+ */
+class JsonEncoderFactory implements JsonEncoderFactoryInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getEncoder(): JsonEncoder
+    {
+        return new JsonEncoder();
+    }
+}