+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Changed the output of the message that will be shown after installing a new plugin.
+- Default is now the highest stable version for a plugin to install.
+### Removed
+- Dependencies to the composer factory
+
 ## [2.2.3] - 2018-09-25
 ### Fixed
 - Memory exhaustion when executing the plugins:available command.